% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cn.farms}
\alias{cn.farms}
\title{cn.farms}
\usage{
cn.farms(filenames, cores = 1, runtype = "bm")
}
\arguments{
\item{filenames}{the absolute filepaths of the CEL files.}

\item{cores}{number of parallel instances.}

\item{runtype}{either ff or bm.}
}
\value{
An instance of \code{\link[Biobase:ExpressionSet-class]{ExpressionSet}}
containing the results of the analysis.
}
\description{
Wrapper for the cn.farms algorithm
}
\examples{
\dontrun{
require('hapmapsnp6')
celDir <- system.file('celFiles', package = 'hapmapsnp6')
filenames <- dir(path = celDir, full.names = TRUE)
cn.farms(filenames = filenames)
}
}
\author{
Djork-Arne Clevert \email{okko@clevert.de} and
Andreas Mitterecker \email{mitterecker@bioinf.jku.at}
}

