\name{getFragmentSet}
\alias{getFragmentSet}
\title{Finds SNPs which belong to one fragment}
\usage{
  getFragmentSet(fragLength)
}
\arguments{
  \item{fragLength}{fragLength}
}
\value{
  windows for fragments
}
\description{
  Finds SNPs which belong to one fragment
}
\author{
  Djork-Arne Clevert \email{okko@clevert.de} and Andreas
  Mitterecker \email{mitterecker@bioinf.jku.at}
}

