\name{getSingleProbeSetSize}
\alias{getSingleProbeSetSize}
\title{Combines data for probeset summarization}
\usage{
  getSingleProbeSetSize(fsetid)
}
\arguments{
  \item{fsetid}{fsetid}
}
\value{
  a Indices whhich are used for probeset summarization
}
\description{
  Combines data for probeset summarization
}
\author{
  Djork-Arne Clevert \email{okko@clevert.de} and Andreas
  Mitterecker \email{mitterecker@bioinf.jku.at}
}

