\name{normAdd}
\alias{normAdd}
\title{Extracts info from the package name}
\usage{
  normAdd(pkgname)
}
\arguments{
  \item{pkgname}{The package name according to the
  bioconductor annotation names.}
}
\value{
  Additional info for save files.
}
\description{
  Extracts info from the package name
}
\author{
  Andreas Mitterecker
}

