\name{summarizeFarmsMethods}
\alias{summarizeFarmsMethods}
\title{Lists methods for possible FARMS summarization...}
\description{Lists methods for possible FARMS summarization}
\details{Possible FARMS summarization}
\value{Returns a data frame with all possible FARMS calls.}
\author{Djork-Arne Clevert \email{okko@clevert.de} and 
Andreas Mitterecker \email{mitterecker@bioinf.jku.at}}
\examples{summarizeFarmsMethods()}

