\name{summarizeFarmsStatistics}
\alias{summarizeFarmsStatistics}
\title{Mean or median instead of the FARMS model}
\usage{
  summarizeFarmsStatistics(probes, type = "median", ...)
}
\arguments{
  \item{probes}{A matrix with numeric values.}

  \item{type}{The statistic which you want to apply.}

  \item{...}{Further parameters}
}
\value{
  Some data
}
\description{
  Mean or median instead of the FARMS model
}
\author{
  Andreas Mitterecker
}

