\name{summarizeWindowMethods}
\alias{summarizeWindowMethods}
\title{Lists methods for possible window methods...}
\description{Lists methods for possible window methods}
\details{Function to list how neighbouring positions can be combined.}
\value{Returns a data frame with all possible methods.}
\author{Djork-Arne Clevert \email{okko@clevert.de} and 
Andreas Mitterecker \email{mitterecker@bioinf.jku.at}}
\examples{summarizeWindowMethods()}

