\name{summarizeFarmsMethods}
\alias{summarizeFarmsMethods}
\title{Lists methods for possible FARMS summarization}
\value{
  Returns a data frame with all possible FARMS calls.
}
\description{
  Possible FARMS summarization
}
\examples{
summarizeFarmsMethods()
}
\author{
  Djork-Arne Clevert \email{okko@clevert.de} and Andreas
  Mitterecker \email{mitterecker@bioinf.jku.at}
}

