\name{summarizeWindowMethods}
\alias{summarizeWindowMethods}
\title{Lists methods for possible window methods}
\value{
  Returns a data frame with all possible methods.
}
\description{
  Function to list how neighbouring positions can be
  combined.
}
\examples{
summarizeWindowMethods()
}
\author{
  Djork-Arne Clevert \email{okko@clevert.de} and Andreas
  Mitterecker \email{mitterecker@bioinf.jku.at}
}

